Pull out fetch calls to helper components
