+++ src/components/Helpers/People.js
@@ -1,0 +1,26 @@
+const People = async () => {
+  const url = `https://swapi.co/api/people/`;
+  const response = await fetch(url);
+  const data = await response.json();
+
+  return Promise.all(data.results.map(async (person, index) => {
+    const currHomeworld = await fetchPropertyObj(person.homeworld);
+    const currSpecies = await fetchPropertyObj(person.species[0]);
+    return {
+      name: person.name,
+      homeworld: currHomeworld.name,
+      population: currHomeworld.population,
+      species: currSpecies.name,
+      index,
+      favorite: false,
+    };
+  }))
+}
+
+const fetchPropertyObj = async (url) => {
+  const response = await fetch(url);
+  const propertyObj = await response.json();
+  return propertyObj;
+}
+
+export default People;
